Add an event messaging system
A basic publish/subscribe event messaging system has been added to
provide a way of communicating between systems. The queue and flush
feature allows for cases when a server would want to receive incoming
network messages without blocking. The server could queue them until
the start of it's next frame where it would then flush all changes so
that it's systems can act on the updates in a logical order.