Add input handling system
Added a command based input system that currently only handles keys.
Eventually this will also handle input from other sources such as mouse.
The system is designed to be called as a pre-simulator so that all input
commands are eecuted before the main system calls.

A consumer should register a command with the system as a
KeyCommandExecutor or KeyCommandExecutorFunc. The delta time will be
passed into all executors so simple callbacks can be run for small
tasks.

In order for an executor to be triggered, a KeyPressedEvent or
KeyReleasedEvent should be dispatched through the given Subscriber.