Add the main engine package
The engine package is responsible for handling the orchestration of
system execution.

The Engine can be configured via the defined Options.

A default game loop is supplied based on a fixed step simulation
similar to deWitters game loop or the fix your timestep article.

The EntityManager provides the consumer with a way to create Entities
within the Engine. An Entity represents a single "thing" within the
simulations and will be passed to systems to group components to that
particular Entity. Essentially an Entity is just an identifier for a
grouping of components.